+++ lib/features/number_trivia/domain/usecases/get_concrete_number_usecase.dart
@@ -20,7 +20,7 @@ class GetConcreteNumberUseCase extends UseCase<NumberTrivia, Params>{
 }
 
 class Params extends Equatable {
-  final int number;
+  final double number;
   const Params({required this.number});
 
   @override
